+++ scaffold/CHANGELOG.md
@@ -25,3 +25,13 @@
 * (2)、[完成build打包模块的开发](https://dobettest.atlassian.net/jira/software/c/projects/WREN/boards/5?modal=detail&selectedIssue=WREN-7)
 ### 二、修复的问题
 * (1)、[env变量注入process.env后类型变化导致的判断错误问题](https://dobettest.atlassian.net/jira/software/c/projects/WREN/boards/5?selectedIssue=WREN-21)
+
+<hr>
+
+## 2、Version(0.1.4)
+
+### 一、新增内容
+* (1)、[optionsCached](https://dobettest.atlassian.net/jira/software/c/projects/WREN/boards/5?selectedIssue=WREN-25)
+### 二、修复的问题
+* (1)、[vue3分包兼容处理](https://dobettest.atlassian.net/jira/software/c/projects/WREN/boards/5?selectedIssue=WREN-23)
+* (2)、[移除copy-webpack-plugin的一些无用资源](https://dobettest.atlassian.net/jira/software/c/projects/WREN/boards/5?modal=detail&selectedIssue=WREN-24)